% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{Drought}
\alias{Drought}
\title{Drought Dataset}
\format{
A data frame with 61828 rows and 12 columns:
\describe{
\item{MapDate}{Date in form YYYYMMDD}
\item{FIPS}{County identifier}
\item{County}{County name}
\item{State}{State abbreviation}
\item{None}{Drought Severity Category 1}
\item{D0}{Drought Severity Category 2}
\item{D1}{Drought Severity Category 3}
\item{D2}{Drought Severity Category 4}
\item{D3}{Drought Severity Category 5}
\item{D4}{Drought Severity Category 6}
\item{ValidEnd}{End Drought Date in YYYY-MM-DD}
\item{ValidStart}{Start Drought Date in YYYY-MM-DD}
}
}
\source{
US Drought Monitor, dataset names for the same object as Drought.rda.
}
\usage{
data(Drought)
}
\description{
Drought is the variable name of the DataFrame in Drought.rda file
}
\keyword{datasets}
