% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_Functions.R
\name{drought_main}
\alias{drought_main}
\title{Main function that visualizes drought data using heatmaps}
\usage{
drought_main(drought_data, start_date, end_date)
}
\arguments{
\item{drought_data}{n by p matrix containing n data points to cluster.}

\item{start_date}{string of format "yyyy-mm-dd", starting date of time series}

\item{end_date}{string of format "yyyy-mm-dd", ending date of time series}
}
\value{
a heatmap plot of the state's county names on x axis, week time series on y axis and heatmap colored by drought severity category
}
\description{
Main function that visualizes drought data using heatmaps
}
\examples{
# Give example
}
