% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{ENSO}
\alias{ENSO}
\title{ENSO Dataset}
\format{
A data frame with 72 rows and 13 columns:
\describe{
\item{Season}{Season is a two year range: YYYY-YYYY}
\item{JJA}{Oceanic Niño Index value (numeric) for June July August }
\item{JAS}{Oceanic Niño Index value (numeric) for July August Sept}
\item{ASO}{Oceanic Niño Index value (numeric) for August Sept Oct }
\item{SON}{Oceanic Niño Index value (numeric) for Sept Oct Nov}
\item{OND}{Oceanic Niño Index value (numeric) for Oct Nov Dec }
\item{NDJ}{Oceanic Niño Index value (numeric) for Nov Dec Jan }
\item{DJF}{Oceanic Niño Index value (numeric) for Dec Jan Feb}
\item{JFM}{Oceanic Niño Index value (numeric) for Jan Feb Mar }
\item{FMA}{Oceanic Niño Index value (numeric) for Feb Mar Apr}
\item{MAM}{Oceanic Niño Index value (numeric) for Mar Apr May }
\item{AMJ}{Oceanic Niño Index value (numeric) for Apr May Jun}
\item{MJJ}{Oceanic Niño Index value (numeric) for May Jun Jul }
}
}
\source{
NOAA
}
\usage{
data(ENSO)
}
\description{
Time series ENSO indices with monthly values.
}
\keyword{datasets}
