% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENSO.R
\name{enso_main}
\alias{enso_main}
\title{Functions that create a ENSO Time Series Plot of ONI values}
\arguments{
\item{dataframe}{n by p matrix containing ONI values, from National Weather Service, Climate Prediction Center: https://origin.cpc.ncep.noaa.gov/products/analysis_monitoring/ensostuff/ONI_v5.php}

\item{StartDate}{string value for season start date "year1-year2", the starting date of your subset, i.e. "2000-2001", comes from season column of the dataframe}

\item{EndDate}{string value for season end date "year19-year20", the starting date of your subset, i.e. "2020-2021", comes from season column of the dataframe}
}
\value{
p Main package plot which is a barplot visualization of the ENSO intensities by month
}
\description{
Functions that create a ENSO Time Series Plot of ONI values
}
\examples{
# Give example
# Load the data
sample_data <- data.frame(
Season = c("2000-2001", "2001-2002"),
JJA = c(-0.6, -0.1),
JAS = c(-0.5, -0.1),
ASO = c(-0.5, -0.2),
SON = c(-0.6, -0.3),
OND = c(-0.7, -0.3),
NDJ = c(-0.7, -0.3),
DJF = c(-0.7, -0.1),
JFM = c(-0.5, 0.0),
FMA = c(-0.4, 0.2),
MAM = c(-0.3, 0.2),
AMJ = c(-0.3, 0.4),
MJJ = c(-0.1, 0.7)
)
print(sample_data)
enso_main(sample_data, "2000-2001", "2001-2002")
}
