% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_Functions.R
\name{SubsetData}
\alias{SubsetData}
\title{Functions that visualize spatiotemporal data using heatmaps}
\usage{
SubsetData(X)
}
\arguments{
\item{X}{n by p matrix containing n data points to cluster.}

\item{K}{An integer specifying number of clusters.}

\item{M}{K by p matrix of cluster centers.}
}
\value{
Explain return
}
\description{
Functions that visualize spatiotemporal data using heatmaps
}
\examples{
# Give example
}
